Fix shader precision mismatch and SVG circle attribute errors
Found by actually running the app in a browser and inspecting the
console (build/test alone can't catch either class of bug):

- Dust fragment shader in HeroScene declared uTime as mediump while
  the vertex shader defaults it to highp, which WebGL's linker
  rejects for a uniform shared across both stages.
- ProofInNumbers' traveling-light circles only set cx/cy via the
  animate keyframes, leaving them undefined on first paint.

+++ src/app/three/HeroScene.tsx
@@ -786,7 +786,12 @@ const DUST_VERTEX_SHADER = /* glsl */ `
   }
 `;
 const DUST_FRAGMENT_SHADER = /* glsl */ `
-  precision mediump float;
+  // highp, not mediump — uTime is also a uniform in DUST_VERTEX_SHADER,
+  // which (per GLSL ES spec) defaults floats to highp with no explicit
+  // precision statement; mediump here previously created a precision
+  // mismatch WebGL rejects at link time ("Precisions of uniform 'uTime'
+  // differ between VERTEX and FRAGMENT shaders").
+  precision highp float;
   uniform float uTime;
   varying float vSeed;
   void main() {
